refactor: relocate "get" utility function
